requirements.txt
python-telegram-bot==20.3
pymongo==4.4.1
requests==2.31.0
Pillow==10.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
